fishing-dashboard-deploy/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
python-multipart==0.0.9
passlib==1.7.4
python-jose[cryptography]==3.3.0
